# speed up `get_model_spec()` helper