--- a/R/translate.R
+++ b/R/translate.R
@@ -106,25 +106,21 @@ get_model_spec <- function(model, mode, engine) {
   env_obj <- grep(model, env_obj, value = TRUE)
 
   res <- list()
-  res$libs <-
-    rlang::env_get(m_env, paste0(model, "_pkgs")) %>%
-    dplyr::filter(engine == !!engine) %>%
-    purrr::pluck("pkg") %>%
-    purrr::pluck(1)
 
-  res$fit <-
-    rlang::env_get(m_env, paste0(model, "_fit")) %>%
-    dplyr::filter(mode == !!mode & engine == !!engine) %>%
-    dplyr::pull(value) %>%
-    purrr::pluck(1)
+  libs <- rlang::env_get(m_env, paste0(model, "_pkgs"))
+  libs <- vctrs::vec_slice(libs$pkg, libs$engine == engine)
+  res$libs <- libs[[1L]]
 
-  pred_code <-
-    rlang::env_get(m_env, paste0(model, "_predict")) %>%
-    dplyr::filter(mode == !!mode & engine == !!engine) %>%
-    dplyr::select(-engine, -mode)
+  fits <- rlang::env_get(m_env, paste0(model, "_fit"))
+  fits <- vctrs::vec_slice(fits$value, fits$mode == mode & fits$engine == engine)
+  res$fit <- fits[[1L]]
 
-  res$pred <- pred_code[["value"]]
-  names(res$pred) <- pred_code$type
+  preds <- rlang::env_get(m_env, paste0(model, "_predict"))
+  where <- preds$mode == mode & preds$engine == engine
+  types <- vctrs::vec_slice(preds$type, where)
+  values <- vctrs::vec_slice(preds$value, where)
+  names(values) <- types
+  res$pred <- values
 
   res
 }