--- a/R/translate.R
+++ b/R/translate.R
@@ -109,11 +109,11 @@ get_model_spec <- function(model, mode, engine) {
 
   libs <- rlang::env_get(m_env, paste0(model, "_pkgs"))
   libs <- vctrs::vec_slice(libs$pkg, libs$engine == engine)
-  res$libs <- libs[[1L]]
+  res$libs <- if (length(libs) > 0) {libs[[1]]} else {NULL}
 
   fits <- rlang::env_get(m_env, paste0(model, "_fit"))
   fits <- vctrs::vec_slice(fits$value, fits$mode == mode & fits$engine == engine)
-  res$fit <- fits[[1L]]
+  res$fit <- if (length(fits) > 0) {fits[[1]]} else {NULL}
 
   preds <- rlang::env_get(m_env, paste0(model, "_predict"))
   where <- preds$mode == mode & preds$engine == engine