@simonpcouch 2023-03-06
That's a good thought! There are no `NEWS` updates re: `vec_slice()` after our minimum `0.4.1`, and the helper runs fine with that version, so I believe we should be good to go.